Conflicts:
	modules/juce_audio_plugin_client/AU/juce_AU_Wrapper.mm

+++ modules/juce_core/memory/juce_MemoryBlock.cpp
@@ -369,7 +369,7 @@ String MemoryBlock::toBase64Encoding() const
 
     String destString ((unsigned int) size); // store the length, followed by a '.', and then the data.
     const int initialLen = destString.length();
-    destString.preallocateBytes (sizeof (String::CharPointerType::CharType) * (initialLen + 2 + numChars));
+    destString.preallocateBytes (sizeof (String::CharPointerType::CharType) * (size_t) (initialLen + 2 + numChars));
 
     String::CharPointerType d (destString.getCharPointer());
     d += initialLen;
